Fixed auto-install for nixos install issue

 On branch main
 Your branch is up to date with 'origin/main'.

 Changes to be committed:
	modified:   auto-install.sh

+++ auto-install.sh
@@ -302,8 +302,8 @@ fi
 #return to NixOS-Hyprland
 cd ~/NixOS-Hyprland
 
-# copy fastfetch config if nixos.png is not present
-if [ ! -f "$HOME/.config/fastfetch/nixos.png" ]; then
+# copy fastfetch config if nixos image is not present
+if [ ! -f "$HOME/.config/fastfetch/nixos.png" ] && [ ! -f "$HOME/.config/fastfetch/images/nixos.png" ]; then
     cp -r assets/fastfetch "$HOME/.config/"
 fi
